Pattern match fix.
Pattern match was done as
  $SKIP_DIR =~ $sub_dir

This used to work because we have lower-cased directories
in pams and a mix-case directory name in StRoot. It never
occured to have a wrong matching. However, when I had to
exclude Stl3RawReaderMaker from the list of makers, an
interresting side effect was

  $SKIP_DIRS = "blabla ... Stl3RawReaderMaker"

  pams/l3 pops up 'l3' as $sub_dir and of course

  $SKIP_DIRS =~ $sub_dir  becomes TRUE since the string
'l3' belong to Stl3RawReaderMaker. That's a big "oups!".
The net result was that pams/l3 was never compiled (since it
was skipped) and consequently, an avalanche of compilation
problems occured (since some of the tables are required not
only by Stl3RawReaderMaker but also by some other base makers).

That's for today's educational message ...