Introducing the baseline year 2006 geometry, which is "the best" Y2005 geo plus the bugfix in the TPC backplane (not too important). Pending a better definition of the SSD from Lilian which we'll have to version and maybe cut new tags of 2004-5 geometries if there is a need.
As agreed, created UPGR01 and UPGR02 (note cautious use of namespace),
which are basically SSD+HFT and IST+HFT. Note that NONE of engineering
detail is available in either case, and less so for the integration
structural elements. We expect to do a lot of development in this area
as new facts and engineering data are provided. This cut will allow us
to proceed  with tracking studies in the near term.

Fixed a comment and added a couple more.