Modified geometry.g so that seperate particle transport cuts can be used in the BEMC and EEMC.  This is needed for the spin/dijet simulation request summarized here:
http://drupal.star.bnl.gov/STAR/starsimrequests/2010/may/26/filtered-dijet-simulation

The simulation request applies filters which select out barrel- and
adjacent-jetpatch triggers.  Thus, the endcap is only important for
trigger response.  Requestors have noted a x2 increase in speed when
EEMC is run with higher tracking cuts than in the BEMC.

We define the Y2009b tag to apply the 10keV default tracking cuts in
the BEMC, and revert the EEMC to the 80kev/1MeV (photon/electron) cuts.