Reimplemented segmentation of PXL sensor to two halves.
In the sensor's local coordinate system the first half is for x<0 and the second
one is for x>0. The notion of inner and outter halves is not critical and in
fact confusing because it depends on the original rotation around the z axis.
For example, two rotations of phi=5 and phi=-175 give us essentially the same
layer but result in swapped inner and outter halves.