Added y2018a geometry tag.
In 2018 we will be running with a single sector (20) with the inner TPC
upgrade.  This upgrade is not planned for physics production.   The iTPC
geometry models will be implemented in the asymptotic geometry tags (y2018x,...)
and the production tags will maintain the old model.  The goal here is to
ensure that any embedding and/or simulation productions do not inadvertantly
pickup the iTPC geometry model.