We have an issue with different sources of particle information, in particular particle decay modes.
We are using pythia8 to decay particles produced in hijing decays, and occassionally
running into a situation where we decay to a state with a particle unknown to
ROOT's TDatabasePDG...

e.g.

D_s+ --> pi+ f0(890)
              --> K+ K-

Rare enough that it didn't make it into TDatabasePDG, but frequent enough that we
see it happen from pythia 8.

Solution for now is to only permit decays where all final state particles are
known to TDatabasePDG.

Error message will be shown when this decision is made.

Job will be brutally terminated if more that 1k consecutive attempts to
decay a particle fail (indicating user closed down the decay modes to only
those we cannot handle)...