Int_t --> int (and etc...), plus ...
When particles are decayed at (or near) rest, we can run into an issue with
differing definitions in the various particle data tables.  If a particle
has more mass in pythia8 than in hijing, then when it decays at rest, pythia8
will not be able to decay it.  This results in the decayer code going into
an infinite loop, trying to decay something that won't decay.

Solution is the following.  When a particle is decayed by pythia8, we take
as given the momentum of the particle.  But we recalculate the energy of the
particle using the mass tabulated by pythia 8.