Proposed fix for ticket 3399-- https://www.star.bnl.gov/rt3/Ticket/Display.html?id=3399&results=8b3b60b210def723798e28371b984406
Relies on changes to the geometry which flag the mixed iTPC/TPC mode.
-- StarGeo.xml v1.29 / TpceConfig.xml v1.7 / TpceGeo3a.xml v1.20

When the mixed iTPC/TPC mode is detected, the outer padrows in sector
20 are renumbered to match the iTPC numbering scheme (41 through 72).

The inner padrows keep the old TPC numbering scheme (1 through 13) with 3
pseudopadrows per real padrow.  To hopefully avoid confusion, we place these
in a fake sector 99.  TpcRS should skip these in the hit processing loop.
We place them in a fake sector 99.  TpcRS should skip these in the hit processing
loop.

Code has been minimally tested as per the recipe in the trouble ticket.
Sector 20 shows reasonable (as many as 32) matched hits in the small sample
generated.