Remove fixed target from the dev2021 setup.
The fixed target is implemented as a separate geometry module, which places
the fixed target and its supports parallel to the vacuum volume of the beam pipe
segments.  This requires an overlap to be flagged.  We are trying to eliminate
overlaps in the dev2021 geometry, as it slows down the track fitting w/ genfit.

Fixed target will require a reshape of the beampipe.